+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/validation/impl/NamedDynamicTemplateListValidator.java
@@ -14,6 +14,7 @@ import java.util.stream.Collectors;
 
 import org.hibernate.search.backend.elasticsearch.lowlevel.index.mapping.impl.DynamicTemplate;
 import org.hibernate.search.backend.elasticsearch.lowlevel.index.mapping.impl.NamedDynamicTemplate;
+import org.hibernate.search.backend.elasticsearch.reporting.impl.ElasticsearchValidationMessages;
 
 public class NamedDynamicTemplateListValidator implements Validator<List<NamedDynamicTemplate>> {
