+++ mapper/javabean/src/main/java/org/hibernate/search/mapper/javabean/reporting/impl/JavaBeanEventContextMessages.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.orm.logging.impl;
+package org.hibernate.search.mapper.javabean.reporting.impl;
 
 import org.hibernate.search.util.common.logging.impl.MessageConstants;
 
@@ -13,16 +13,17 @@ import org.jboss.logging.annotations.Message;
 import org.jboss.logging.annotations.MessageBundle;
 
 /**
- * Message bundle for event contexts in the Hibernate ORM mapper.
+ * Message bundle for event contexts in the JavaBean mapper.
  */
 @MessageBundle(projectCode = MessageConstants.PROJECT_CODE)
-public interface HibernateOrmEventContextMessages {
+public interface JavaBeanEventContextMessages {
 
-	HibernateOrmEventContextMessages INSTANCE = Messages.getBundle( HibernateOrmEventContextMessages.class );
+	JavaBeanEventContextMessages INSTANCE = Messages.getBundle( JavaBeanEventContextMessages.class );
 
-	@Message(value = "Hibernate ORM mapping")
+	@Message(value = "JavaBean mapping")
 	String mapping();
 
 	@Message(value = "Schema management")
 	String schemaManagement();
+
 }
