+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/reporting/impl/HibernateOrmEventContextMessages.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.javabean.log.impl;
+package org.hibernate.search.mapper.orm.reporting.impl;
 
 import org.hibernate.search.util.common.logging.impl.MessageConstants;
 
@@ -16,14 +16,13 @@ import org.jboss.logging.annotations.MessageBundle;
  * Message bundle for event contexts in the JavaBean mapper.
  */
 @MessageBundle(projectCode = MessageConstants.PROJECT_CODE)
-public interface JavaBeanEventContextMessages {
+public interface HibernateOrmEventContextMessages {
 
-	JavaBeanEventContextMessages INSTANCE = Messages.getBundle( JavaBeanEventContextMessages.class );
+	HibernateOrmEventContextMessages INSTANCE = Messages.getBundle( HibernateOrmEventContextMessages.class );
 
-	@Message(value = "JavaBean mapping")
+	@Message(value = "Hibernate ORM mapping")
 	String mapping();
 
 	@Message(value = "Schema management")
 	String schemaManagement();
-
 }
